requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
celery
redis
python-multipart
ffmpeg-python
